Move misc. things to services
