+++ tests/test_tasks/test_prune_inactive_reservations.py
@@ -6,7 +6,7 @@ import pytest
 
 from tilavarauspalvelu.enums import ReservationStateChoice
 from tilavarauspalvelu.models import Reservation
-from tilavarauspalvelu.utils.pruning import prune_inactive_reservations
+from tilavarauspalvelu.services.pruning import prune_inactive_reservations
 from utils.date_utils import local_datetime
 
 from tests.factories import ReservationFactory
