+++ tests/test_tasks/test_prune_recurring_reservations.py
@@ -6,7 +6,7 @@ import pytest
 from freezegun import freeze_time
 
 from tilavarauspalvelu.models import RecurringReservation
-from tilavarauspalvelu.utils.pruning import prune_recurring_reservations
+from tilavarauspalvelu.services.pruning import prune_recurring_reservations
 from utils.date_utils import local_datetime
 
 from tests.factories import RecurringReservationFactory
