+++ tests/test_tasks/test_prune_reservation_with_inactive_payments.py
@@ -8,7 +8,7 @@ from freezegun import freeze_time
 
 from tilavarauspalvelu.enums import OrderStatus, ReservationStateChoice
 from tilavarauspalvelu.models import Reservation
-from tilavarauspalvelu.utils.pruning import prune_reservation_with_inactive_payments
+from tilavarauspalvelu.services.pruning import prune_reservation_with_inactive_payments
 from utils.date_utils import local_datetime
 
 from tests.factories import PaymentOrderFactory, ReservationFactory
