+++ tilavarauspalvelu/models/user/model.py
@@ -14,7 +14,7 @@ from helusers.models import AbstractUser
 from tilavarauspalvelu.enums import ReservationNotification, UserRoleChoice
 from tilavarauspalvelu.integrations.helauth.typing import IDToken
 from tilavarauspalvelu.integrations.helauth.utils import get_jwt_payload
-from tilavarauspalvelu.utils.permission_resolver import PermissionResolver
+from tilavarauspalvelu.services.permission_resolver import PermissionResolver
 from utils.date_utils import DEFAULT_TIMEZONE
 
 from .queryset import ProfileUserManager, UserManager
