--- tilavarauspalvelu/utils/image_purge.py
@@ -1,1 +1,0 @@
-
